Finished restructuring
