+++ server/api/index.js
@@ -5,4 +5,6 @@ const router = new Router();
 router.use('/assignments', require('./assignments'));
 router.use('/setup', require('./setup'));
 router.use('/integrations', require('./integrations'));
-router.user('/students', require('./students'));
+router.use('/students', require('./students'));
+
+module.exports = router.routes();
