+++ server/api/integrations/index.js
@@ -1,7 +1,7 @@
 const Router = require('koa-router');
 const router = new Router();
 
-const Ctrl = require('../../controllers/api/integrations');
+const Ctrl = require('./integrations.controller');
 
 router.post('/sms/submit', Ctrl.submitSMS);
 router.post('/sms/verify', Ctrl.verifySMS);
